+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Exclude empty `virtual` and `override` methods in `EmptyMethodRule` if they have comments.
+
 ## [0.5.0] - 2019-08-09
 
 ### Added