+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Exclude empty `virtual` and `override` methods in `EmptyMethodRule` if they have comments.
+- Include constructors and destructors in `NoSemiAfterMethodDeclarationRule`.
 
 ## [0.5.0] - 2019-08-09