+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `CommentRegexRule` analysis rule template, which flags comments matching a regular expression.
+
 ### Changed
 
 - Exclude empty `virtual` and `override` methods in `EmptyMethodRule` if they have comments.
 - Include constructors and destructors in `NoSemiAfterMethodDeclarationRule`.
+- Allow `TFrame` prefix in `ClassNameRule`.
 
 ## [0.5.0] - 2019-08-09